+++ src/bin/chrome-devtools-mcp.ts
@@ -6,6 +6,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+process.title = 'chrome-devtools-mcp';
+
 import {version} from 'node:process';
 
 const [major, minor] = version.substring(1).split('.').map(Number);